added update floor map item route

+++ floor-map-service/src/main/java/com/jake/waitlistservice/jpa/repository/FloorMapItemRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface FloorMapItemRepository extends JpaRepository<FloorMapItem, BigInteger> {
+public interface FloorMapItemRepository extends JpaRepository<FloorMapItem, Long> {
     List<FloorMapItem> findAllByFloorMapId(long floorMapId, Sort sortedBy);
 
     List<FloorMapItem> findAllByFloorMapId(long floorMapId);
